Fixed possible error for Workshop update endpoint while converting null for available seats (#1311)
* Fixed possible error for Workshop update endpoint while converting null available seats

* Fixed tests

* Fixed workshop creation

* Fixed unit tests

* Added tests for creation and update

+++ OutOfSchool/OutOfSchool.WebApi.Tests/Controllers/ApplicationControllerTests.cs
@@ -852,7 +852,7 @@ public class ApplicationControllerTests
             InstitutionHierarchyId = w.InstitutionHierarchyId,
             InstitutionId = w.InstitutionId,
             Institution = w.Institution,
-            AvailableSeats = w.AvailableSeats,
+            AvailableSeats = (uint)w.AvailableSeats,
             TakenSeats = w.TakenSeats,
         }).ToList();
     }
